requirements.txt
Django==1.9.8
certifi==2016.2.28
dj-database-url==0.4.1
future==0.15.2
gunicorn==19.6.0
lxml==3.6.1
psycopg2==2.6.2
python-telegram-bot==5.0.0
redis==2.10.5
requests==2.10.0
urllib3==1.16
wsgiref==0.1.2
